+++ praktikumsplaner-frontend/frontend/src/types/Ausbildungsjahr.ts
@@ -1,6 +1,6 @@
 export const Ausbildungsjahr = [
     {
-        name: "ab 2. Lehrjahr",
+        name: "2. Lehrjahr",
         value: "JAHR2",
         zeitraumFISI: "01.09 - 28.02 oder 01.03 - 31.08",
     },
