+++ praktikumsplaner-frontend/frontend/src/views/Praktikumsplaetze/MeldungAusbildung.vue
@@ -217,72 +217,12 @@
                     </v-col>
                     <v-col cols="2" />
                     <v-col>
-                        <v-select
-                            v-model="praktikumsstelle.ausbildungsjahr"
-                            label="Ausbildungsjahr*"
-                            :items="Ausbildungsjahr"
-                            item-value="value"
-                            item-text="name"
-                            :rules="requiredArrayRule"
-                            :menu-props="customMenuProps"
-                            outlined
-                            multiple
-                            @change="
-                                () => {
-                                    sortAusbildungsjahre();
-                                }
+                        <SelectMultipleAusbildungsjahr
+                            :praktikumsstelle="praktikumsstelle"
+                            @update:praktikumsstelle="
+                                (value) => (praktikumsstelle = value)
                             "
-                        >
-                            <template #prepend-item>
-                                <v-list-item
-                                    ripple
-                                    @mousedown.prevent
-                                    @click="selectAllAusbildungsjahre"
-                                >
-                                    <v-list-item-action>
-                                        <v-icon
-                                            :color="ausbildungsjahrIconColor()"
-                                        >
-                                            {{ AusbildungsjahrIcon }}
-                                        </v-icon>
-                                    </v-list-item-action>
-                                    <v-list-item-content>
-                                        <v-list-item-title>
-                                            Egal
-                                        </v-list-item-title>
-                                    </v-list-item-content>
-                                </v-list-item>
-                                <v-divider class="mt-2"></v-divider>
-                            </template>
-                            <template #item="data">
-                                <v-list-item-action>
-                                    <v-icon
-                                        v-if="
-                                            data.attrs['aria-selected'] ===
-                                            'true'
-                                        "
-                                    >
-                                        mdi-checkbox-marked
-                                    </v-icon>
-                                    <v-icon v-else>
-                                        mdi-checkbox-blank-outline
-                                    </v-icon>
-                                </v-list-item-action>
-                                <v-list-item-content>
-                                    <v-list-item-title>
-                                        {{ data.item.name }}
-                                    </v-list-item-title>
-                                    <v-list-item-subtitle
-                                        v-if="
-                                            praktikumsstelle.ausbildungsrichtung ===
-                                            'FISI'
-                                        "
-                                    >
-                                        {{ data.item.zeitraumFISI }}
-                                    </v-list-item-subtitle>
-                                </v-list-item-content>
-                            </template>
-                        </v-select>
+                        />
                     </v-col>
                     <v-col cols="1" />
                 </v-row>
@@ -390,10 +330,9 @@
 
 <script setup lang="ts">
 import { Ausbildungsrichtung } from "@/types/Ausbildungsrichtung";
-import { onMounted, ref, computed } from "vue";
+import { computed, onMounted, ref } from "vue";
 import Praktikumsstelle from "@/types/Praktikumsstelle";
 import { useRules } from "@/composables/rules";
-import { Ausbildungsjahr } from "@/types/Ausbildungsjahr";
 import { Referat } from "@/types/Referat";
 import { YesNoBoolean } from "@/types/YesNoBoolean";
 import { Dringlichkeit } from "@/types/Dringlichkeit";
@@ -406,6 +345,7 @@ import { useFormatter } from "@/composables/formatter";
 import { APP_SECURITY } from "@/Constants";
 import { useUserStore } from "@/stores/user";
 import Meldezeitraum from "@/types/Meldezeitraum";
+import SelectMultipleAusbildungsjahr from "@/components/Praktikumsstellen/SelectMultipleAusbildungsjahr.vue";
 
 const activeMeldezeitraum = ref<boolean>(false);
 
@@ -417,13 +357,10 @@ const userStore = useUserStore();
 const validationRules = useRules();
 const formatter = useFormatter();
 const requiredRule = [validationRules.notEmptyRule("Darf nicht leer sein.")];
-const requiredArrayRule = [
-    validationRules.notEmptyArrayRule("Darf nicht leer sein."),
-];
 const emailRule = [
     validationRules.notEmptyRule("Darf nicht leer sein."),
     validationRules.regexRule(
-        /^[A-Z0-9._%+-]{1,64}@[A-Z0-9.-]{1,63}\.[A-Z]{1,63}$/,
+        /^[A-Za-z0-9._%+-]{1,64}@[A-Za-z0-9.-]{1,63}\.[A-Za-z]{1,63}$/,
         "Keine gültige Email."
     ),
     validationRules.maxLengthRule(
@@ -476,27 +413,6 @@ const meldezeitraeume = computed(() => {
 const currentMeldezeitraum = ref<Meldezeitraum>();
 const upcomingMeldezeitraeume = ref<Meldezeitraum[]>([]);
 const passedMeldezeitraeume = ref<Meldezeitraum[]>([]);
-
-const allAusbildungsjahreSelected = computed(() => {
-    return (
-        praktikumsstelle.value.ausbildungsjahr?.length ===
-        Ausbildungsjahr.length
-    );
-});
-
-const someAusbildungsjahreSelected = computed(() => {
-    return (
-        praktikumsstelle.value.ausbildungsjahr?.length !== 0 &&
-        praktikumsstelle.value.ausbildungsjahr?.length !== undefined &&
-        !allAusbildungsjahreSelected.value
-    );
-});
-
-const AusbildungsjahrIcon = computed(() => {
-    if (allAusbildungsjahreSelected.value) return "mdi-checkbox-marked";
-    if (someAusbildungsjahreSelected.value) return "mdi-minus-box";
-    return "mdi-checkbox-blank-outline";
-});
 
 onMounted(() => {
     MeldezeitraumService.getCurrentMeldezeitraum()
@@ -551,27 +467,6 @@ function uploadPraktikumsstelle() {
             resetForm();
         });
     }
-}
-
-function selectAllAusbildungsjahre() {
-    if (allAusbildungsjahreSelected.value) {
-        praktikumsstelle.value.ausbildungsjahr = [];
-    } else {
-        praktikumsstelle.value.ausbildungsjahr = Ausbildungsjahr.map(
-            (ausbildungsjahr) => ausbildungsjahr.value
-        );
-    }
-}
-
-function sortAusbildungsjahre() {
-    praktikumsstelle.value.ausbildungsjahr?.sort((a, b) => a.localeCompare(b));
-}
-
-function ausbildungsjahrIconColor() {
-    return allAusbildungsjahreSelected.value ||
-        someAusbildungsjahreSelected.value
-        ? "primary"
-        : "darkgrey";
 }
 </script>
 <style lang="scss">
